Cost optimization with spot instances

+++ template/ec2_install_apache.sh
@@ -49,15 +49,21 @@ wget https://github.com/derailed/k9s/releases/download/v0.32.5/k9s_linux_arm64.d
 sudo dpkg --install k9s_linux_arm64.deb
 rm -f k9s_linux_arm64.deb
 
-# Add Docker's official GPG key
-sudo apt-get update
-sudo apt-get install -y ca-certificates curl
-sudo install -m 0755 -d /etc/apt/keyrings
-sudo curl -fsSL https://download.docker.com/linux/debian/gpg -o /etc/apt/keyrings/docker.asc
-sudo chmod a+r /etc/apt/keyrings/docker.asc
+if [ -f /etc/debian_version ]; then
+    # Add Docker's official GPG key if Debian
+    sudo apt-get update
+    sudo apt-get install -y ca-certificates curl
+    sudo install -m 0755 -d /etc/apt/keyrings
+    sudo curl -fsSL https://download.docker.com/linux/debian/gpg -o /etc/apt/keyrings/docker.asc
+    sudo chmod a+r /etc/apt/keyrings/docker.asc
 
-# Add Docker repository to Apt sources
-echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+    # Add Docker repository to Apt sources
+    echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+elif [ -f /etc/lsb-release ]; then
+    # Install Docker if it's Ubuntu
+    sudo apt-get update
+    sudo snap install docker
+fi
 
 sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
 # Update package list
